build(cmake): Generate bins to dedicated directory
Binaries will now be generated in their own dedicated directory, to save
having to locate them. A statement has been removed that was making a
build directory within the build directory; the user should just be
prompted to make a build directory in the README, as it's a very simple
task.